refactor!: collapse Z*/UBID* LDAP class hierarchy (#1036)
* refactor: directly instantiate UBIDSearchControls

* refactor: ZLdap not abstract

- there is no point to make an abstraction if then the code casts it. ZLdapContext is strictly tied to UBID and LDAP. This removes also the need to use singletons

* refactor: more UBID -> ZLDAP refactor

* refactor: more UBID -> ZLDAP refactor pt.2

* refactor: more UBID -> ZLDAP refactor pt.3

* refactor: remove unused ILdapGalContext.java

* refactor: make LdapClient instance private

* test: use new zldap classes instead of ubid

* chore: remove unused method

 BREAKING CHANGE: removed public classes from
  com.zimbra.cs.ldap.unboundid: UBIDLdapClient, UBIDLdapContext,
  UBIDAttributes, UBIDMutableEntry, UBIDLdapFilter, UBIDLdapSchema,
  UBIDLdapException, UBIDModificationList, UBIDSearchControls,
  UBIDSearchResultEntry, UBIDSearchResultEnumeration, UBIDSearchScope,
  UBIDLdapFilterFactory. Use the Z* counterparts in
  com.zimbra.cs.ldap. Some package-private members on remaining
  unboundid helpers are now public.
