--- store/src/main/java/com/zimbra/cs/ldap/ILdapGalContext.java
@@ -1,10 +1,0 @@
-// SPDX-FileCopyrightText: 2022 Synacor, Inc.
-// SPDX-FileCopyrightText: 2022 Zextras <https://www.zextras.com>
-//
-// SPDX-License-Identifier: GPL-2.0-only
-
-package com.zimbra.cs.ldap;
-
-public interface ILdapGalContext {
-    //TODO
-}
