+++ store/src/main/java/com/zimbra/cs/ldap/ZAttributes.java
@@ -5,38 +5,54 @@
 
 package com.zimbra.cs.ldap;
 
+import com.unboundid.ldap.sdk.Attribute;
+import com.unboundid.ldap.sdk.Entry;
+import com.unboundid.ldap.sdk.SearchResultEntry;
+import com.zimbra.common.mailbox.ContactConstants;
+import com.zimbra.common.util.ByteUtil;
+import com.zimbra.cs.account.AttributeManager;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import com.zimbra.cs.account.AttributeManager;
+public class ZAttributes extends ZLdapElement implements IAttributes {
 
-public abstract class ZAttributes extends ZLdapElement implements IAttributes {
-    
-    /**
-     * exception for getAttrs(),
-     * 
-     * - If a method does not have a CheckBinary parameter, it will *not* check 
-     *   for binary data and binary transfer based on AttributeManager.
-     *   It will assume all attributes are *not* binary.
-     *   
-     * - If a method has a CheckBinary parameter, it will check for binary data 
-     *   and binary transfer based on AttributeManager if CheckBinary is CHECK.
-     *   It will assume all attributes are *not* binary if CheckBinary is NOCHECK.
-     */
-    
+    private static String[] EMPTY_STRING_ARRAY = new String[0];
+
+    //
+    // The wrapped object here is actually the SearchResultEntry or Entry object.
+    // Unlike JNDI, unboundid handles attributes on the (SearchResult)Entry object.
+    //
+    private Entry entry;
+
+    public ZAttributes(SearchResultEntry entry) {
+        this.entry = entry;
+    }
+
+    public ZAttributes(Entry entry) {
+        this.entry = entry;
+    }
+
+    @Override
+    public void debug() {
+        for (Attribute attr : entry.getAttributes()) {
+            println(attr.toString());
+        }
+    }
+
     @Override
     public String getAttrString(String attrName) throws LdapException {
         return getAttrString(attrName, CheckBinary.NOCHECK);
     }
-    
+
     // make public if necessary
-    private String getAttrString(String attrName, CheckBinary checkBinary) 
+    private String getAttrString(String attrName, CheckBinary checkBinary)
     throws LdapException {
         boolean containsBinaryData;
         String transferAttrName;
-        
+
         if (checkBinary == CheckBinary.NOCHECK) {
             containsBinaryData = false;
             transferAttrName = attrName;
@@ -44,24 +60,24 @@ public abstract class ZAttributes extends ZLdapElement implements IAttributes {
             AttributeManager attrMgr = AttributeManager.getInst();
             containsBinaryData = attrMgr == null ? false : attrMgr.containsBinaryData(attrName);
             boolean isBinaryTransfer = attrMgr == null ? false : attrMgr.isBinaryTransfer(attrName);
-            
+
             transferAttrName = LdapUtil.attrNameToBinaryTransferAttrName(isBinaryTransfer, attrName);
         }
-        
+
         return getAttrString(transferAttrName, containsBinaryData);
     }
-    
+
     @Override
     public String[] getMultiAttrString(String attrName) throws LdapException {
         return getMultiAttrString(attrName, CheckBinary.NOCHECK);
     }
-    
+
     // make public if necessary
-    private String[] getMultiAttrString(String attrName, CheckBinary checkBinary) 
+    private String[] getMultiAttrString(String attrName, CheckBinary checkBinary)
     throws LdapException {
         boolean containsBinaryData;
         boolean isBinaryTransfer;
-        
+
         if (checkBinary == CheckBinary.NOCHECK) {
             containsBinaryData = false;
             isBinaryTransfer = false;
@@ -70,19 +86,19 @@ public abstract class ZAttributes extends ZLdapElement implements IAttributes {
             containsBinaryData = attrMgr == null ? false : attrMgr.containsBinaryData(attrName);
             isBinaryTransfer = attrMgr == null ? false : attrMgr.isBinaryTransfer(attrName);
         }
-        
+
         return getMultiAttrString(attrName, containsBinaryData, isBinaryTransfer);
     }
-    
+
     @Override
-    public String[] getMultiAttrString(String attrName, boolean containsBinaryData, boolean isBinaryTransfer) 
+    public String[] getMultiAttrString(String attrName, boolean containsBinaryData, boolean isBinaryTransfer)
     throws LdapException {
         String transferAttrName = LdapUtil.attrNameToBinaryTransferAttrName(isBinaryTransfer, attrName);
         return getMultiAttrString(transferAttrName, containsBinaryData);
     }
-    
+
     @Override
-    public List<String> getMultiAttrStringAsList(String attrName, CheckBinary checkBinary) 
+    public List<String> getMultiAttrStringAsList(String attrName, CheckBinary checkBinary)
     throws LdapException {
         if (checkBinary == CheckBinary.NOCHECK) {
             return Arrays.asList(getMultiAttrString(attrName, false));
@@ -90,35 +106,105 @@ public abstract class ZAttributes extends ZLdapElement implements IAttributes {
             return Arrays.asList(getMultiAttrString(attrName));
         }
     }
-    
+
     /**
-     * Enumerates over the specified attributes and populates the specified map. 
-     * The key in the map is the attribute ID. For attrs with a single value, 
-     * the value is a String, and for attrs with multiple values the value is an 
+     * Enumerates over the specified attributes and populates the specified map.
+     * The key in the map is the attribute ID. For attrs with a single value,
+     * the value is a String, and for attrs with multiple values the value is an
      * array of Strings.
-     * 
+     *
      * Note: this method always *check* binary.
      */
     public Map<String, Object> getAttrs() throws LdapException {
         return getAttrs(null);
     }
-    
+
+    private String getAttrStringInternal(Attribute attr, boolean containsBinaryData) {
+        if (containsBinaryData) {
+            byte[] bytes = attr.getValueByteArray();
+            return ByteUtil.encodeLDAPBase64(bytes);
+        } else {
+            return attr.getValue();
+        }
+    }
+
+    private String[] getMultiAttrStringInternal(Attribute attr, boolean containsBinaryData) {
+        String[] result = new String[attr.size()];
+
+        if (containsBinaryData) {
+            byte[][] bytesArrays = attr.getValueByteArrays();
+            for (int i = 0; i < bytesArrays.length; i++) {
+                result[i] = ByteUtil.encodeLDAPBase64(bytesArrays[i]);
+            }
+        } else {
+            String[] values = attr.getValues();
+            System.arraycopy(values, 0, result, 0, values.length);
+        }
+        return result;
+    }
+
+    protected String getAttrString(String transferAttrName, boolean containsBinaryData)
+    throws LdapException {
+        Attribute attr = entry.getAttribute(transferAttrName);
+        if (attr != null) {
+            return getAttrStringInternal(attr, containsBinaryData);
+        } else {
+            return null;
+        }
+    }
+
+    protected String[] getMultiAttrString(String transferAttrName, boolean containsBinaryData)
+    throws LdapException {
+        Attribute attr = entry.getAttribute(transferAttrName);
+        // (ZCS-1047) AD sends 'userCertificate;binary' attribute as 'userCertificate' without appending ';binary' to it
+        if (attr == null && transferAttrName.startsWith(ContactConstants.A_userCertificate)) {
+            attr = entry.getAttribute(ContactConstants.A_userCertificate);
+        }
+
+        if (attr != null) {
+            return getMultiAttrStringInternal(attr, containsBinaryData);
+        } else {
+            return EMPTY_STRING_ARRAY;
+        }
+    }
+
     /**
-     * extraBinaryAttrs: if not null, attrs in the set are treated as binary attrs, in addition to 
+     * extraBinaryAttrs: if not null, attrs in the set are treated as binary attrs, in addition to
      * those marked binary in Zimbra's AttributeManager.
      */
-    public abstract Map<String, Object> getAttrs(Set<String> extraBinaryAttrs) throws LdapException;
-    
+    public Map<String, Object> getAttrs(Set<String> extraBinaryAttrs)
+            throws LdapException {
+        Map<String, Object> map = new HashMap<>();
 
-    
-    /**
-     * Retrieves the value for this attribute as a string. 
-     * If this attribute has multiple values, then the first value will be returned.
-     */
-    protected abstract String getAttrString(String transferAttrName, boolean containsBinaryData) 
-    throws LdapException;
-    
-    protected abstract String[] getMultiAttrString(String transferAttrName, boolean containsBinaryData) 
-    throws LdapException;
-    
+        AttributeManager attrMgr = AttributeManager.getInst();
+
+        for (Attribute attr : entry.getAttributes()) {
+            String transferAttrName = attr.getName();
+
+            String attrName = LdapUtil.binaryTransferAttrNameToAttrName(transferAttrName);
+
+            boolean containsBinaryData =
+                (attrMgr != null && attrMgr.containsBinaryData(attrName)) ||
+                (extraBinaryAttrs != null && extraBinaryAttrs.contains(attrName));
+
+            if (attr.size() == 1) {
+                map.put(attrName, getAttrStringInternal(attr, containsBinaryData));
+            } else {
+                String result[] = getMultiAttrStringInternal(attr, containsBinaryData);
+                map.put(attrName, result);
+            }
+        }
+        return map;
+    }
+
+    @Override
+    public boolean hasAttribute(String attrName) {
+        return entry.hasAttribute(attrName);
+    }
+
+    @Override
+    public boolean hasAttributeValue(String attrName, String value) {
+        return entry.hasAttributeValue(attrName, value);
+    }
+
 }
