+++ store/src/main/java/com/zimbra/cs/ldap/ZMutableEntry.java
@@ -5,47 +5,111 @@
 
 package com.zimbra.cs.ldap;
 
+import com.unboundid.asn1.ASN1OctetString;
+import com.unboundid.ldap.sdk.Attribute;
+import com.unboundid.ldap.sdk.Entry;
+import com.zimbra.cs.account.AttributeManager;
+import com.zimbra.cs.ldap.unboundid.UBIDUtil;
+import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
 
 /**
  * Represents a transient, mutable entry in memory.
- *  
- * @author pshao
  */
-public abstract class ZMutableEntry extends ZEntry {
+public class ZMutableEntry extends ZEntry {
 
-    /**
-     * Sets the provided attribute to this entry. 
-     * If this entry already contains an attribute with the same name, 
-     * then the current value will be *replaced* by the new value.
-     */
-    public abstract void setAttr(String attrName, String value);
-    
-    /**
-     * Adds the provided attribute to this entry. 
-     * If this entry already contains an attribute with the same name, 
-     * then their values will be *merged*.
-     */
-    public abstract void addAttr(String attrName, Set<String> values);
-    
-    /**
-     * Adds the provided attribute to this entry. 
-     * If this entry already contains an attribute with the same name, 
-     * then their values will be *merged*.
-     */
-    public abstract void addAttr(String attrName, String value);
-    
-    public abstract String getAttrString(String attrName) throws LdapException;
-    
-    public abstract boolean hasAttribute(String attrName);
+    private Entry entry;
 
-    
-    /**
-     * take a map (key = String, value = String | String[]) and populate ZAttributes.
-     */
-    public abstract void mapToAttrs(Map<String, Object> mapAttrs);
-    
-    public abstract void setDN(String dn);
-    
+    public ZMutableEntry() {
+        // give a bogus dn ("") since unboundid requires it
+        // call site should set a real DN later
+        this.entry = new Entry("");
+    }
+
+    @Override
+    public void debug(ZLdapElementDebugListener debugListener) {
+        print(debugListener, entry.toString());
+    }
+
+    public Entry getNative() {
+        return entry;
+    }
+
+    @Override
+    public ZAttributes getAttributes() {
+        return new ZAttributes(entry);
+    }
+
+    @Override
+    public String getDN() {
+        return entry.getDN();
+    }
+
+    public void setAttr(String attrName, String value) {
+        if (hasAttribute(attrName)) {
+            entry.removeAttribute(attrName);
+        }
+        entry.addAttribute(attrName, value);
+    }
+
+    public void addAttr(String attrName, Set<String> values) {
+        for (String value : values) {
+            entry.addAttribute(attrName, value);
+        }
+    }
+
+    public void addAttr(String attrName, String value) {
+        entry.addAttribute(attrName, value);
+    }
+
+    public String getAttrString(String attrName) throws LdapException {
+        return entry.getAttributeValue(attrName);
+    }
+
+    public boolean hasAttribute(String attrName) {
+        return entry.hasAttribute(attrName);
+    }
+
+    public void mapToAttrs(Map<String, Object> mapAttrs) {
+        AttributeManager attrMgr = AttributeManager.getInst();
+
+        for (Map.Entry<String, Object> me : mapAttrs.entrySet()) {
+
+            String attrName = me.getKey();
+            Object v = me.getValue();
+
+            boolean containsBinaryData = attrMgr == null ? false : attrMgr.containsBinaryData(attrName);
+            boolean isBinaryTransfer = attrMgr == null ? false : attrMgr.isBinaryTransfer(attrName);
+
+            if (v instanceof String) {
+                ASN1OctetString value = UBIDUtil.newASN1OctetString(containsBinaryData, (String) v);
+                Attribute a = UBIDUtil.newAttribute(isBinaryTransfer, attrName, value);
+                entry.addAttribute(a);
+            } else if (v instanceof String[]) {
+                String[] sa = (String[]) v;
+                ASN1OctetString[] values = new ASN1OctetString[sa.length];
+                for (int i = 0; i < sa.length; i++) {
+                    values[i] = UBIDUtil.newASN1OctetString(containsBinaryData, sa[i]);
+                }
+                Attribute a = UBIDUtil.newAttribute(isBinaryTransfer, attrName, values);
+                entry.addAttribute(a);
+            } else if (v instanceof Collection) {
+                Collection c = (Collection) v;
+                ASN1OctetString[] values = new ASN1OctetString[c.size()];
+                int i = 0;
+                for (Object o : c) {
+                    values[i] = UBIDUtil.newASN1OctetString(containsBinaryData, o.toString());
+                    i++;
+                }
+                Attribute a = UBIDUtil.newAttribute(isBinaryTransfer, attrName, values);
+                entry.addAttribute(a);
+            }
+        }
+    }
+
+    public void setDN(String dn) {
+        entry.setDN(dn);
+    }
+
 }
