+++ store/src/main/java/com/zimbra/cs/ldap/unboundid/LdapSSLUtil.java
@@ -74,14 +74,14 @@ public class LdapSSLUtil {
         }
     }
     
-    static SSLContext createSSLContext(boolean allowUntrustedCerts) throws LdapException {
+    public static SSLContext createSSLContext(boolean allowUntrustedCerts) throws LdapException {
         TrustManager tm = getTrustManager(allowUntrustedCerts);
         SSLUtil sslUtil = new SSLUtil(tm);
         
         try {
             return sslUtil.createSSLContext();
         } catch (GeneralSecurityException e) {
-            throw UBIDLdapException.mapToLdapException(e);
+            throw LdapException.mapToLdapException(e);
         }
     }
 }
